feature branch - moving data processing to backend, tidying up routes
